add more api authentication for password

+++ src/ShipperStation.Domain/Entities/Package.cs
@@ -7,7 +7,7 @@ public class Package : BaseAuditableEntity<Guid>
 {
     public string? Name { get; set; }
     public string? Description { get; set; }
-    public decimal PackagePrice { get; set; }
+    public double PackagePrice { get; set; }
     public string? Barcode { get; set; }
     public string ReceiverName { get; set; } = default!;
     public string ReceiverPhone { get; set; } = default!;
